enable linux kernel "make oldocnfig" to accept "y" for new config
When Linux kernel config adds a new option, we will be asked if we
want to set that option to "y/n/m".  Since we want to the script to
automatically build SOS kernel. Just enabel "yes" for all new kconfig
options.  But for product deployment, we need to reduce "yes" for some
of those options. Anyway, this script is for test only.

Signed-off-by: Chaohong guo <chaohong.guo@intel.com>