Initial scripts to read bitcoin blk*.dat file
I plan to use the scripts to teach my daughter of 12-year old and her
fridends about programming in python.  Hopefully, they can understand
and improve it.

Currently the script can read block of version 1, not for verion 2 or
above.  The .dat files are shorten to reduce the size of files for
hosting in github.